Gradle: update Bnd plugin to v6.4.0.
Bnd v6 introduces a `bundle` extension to the the Java plugin's `jar`
task, which no longer relies on conventions (which are deprecated). This
mostly just clears up a deprecation warning. One warning still remains
as long as the Bnd plugin is applied, and there doesn't seem to be
anything we can do about it.